Add latest updated
This was an addendum to the whitepaper, Moving To The Cloud, that I crowdsourced and was finally published v1.0 by the OMG in 2011.  This substantially revised slide deck focuses on proven methods from SEI to evaluate software architectures as a prerequisite to implementation. This spike sprint using these proven methods will clarify options for migration of enterprise applications to the cloud.