Business Computer Network (BCN) 2018-10-28
BCN had this concept to provide free software (ala AOL), free modem (yes free), to any and all
PC enthusiasts and hopefully professionals (B2B). The idea was to index (spider) vast online information services (content/data) and provide a means to search (search engine) the index and make it very easy to access a wide array of information (web). We used a small IBM mainframe for the index, a DEC Vax VMS for search and PDP11/70 UNIX for connection backend services. PC's were new, so there were many flavors, such as CP/M, MS-DOS, DR-DOS, PC-DOS, Apple IIe and this real weird ones called Lisa and Mac that had pictures on the screen that you pointed at with a rolling ball on a wire. We built the very first communications software for the Mac using bootleg programmers reference docs, and it was cross platform. We made it cross platform by designing a common viewing framework (browser) that would interpret a markup language (HTML/XML). I was coding in 'C' and half dozen six assembly languages in a conference room that had a very large table with a assortment of PC's, "the lab". I slept under the conference table many nights even though I lived 2/10 mile away.
"SuperScout" was launched at COMDEX as I stood alongside our company spokesmen, astronauts Buzz Aldrin and Wally Shirra, who were sure to be a draw for the nerds. We had the finest collateral that money could print. Alas the way to monetize these bold feats of technology was to resell the access at slim margins, and because it was so efficient would cannibalize the partners. Since the investment was so high, we would do this at a loss but make up for it on volume. I should've stayed awake in Accounting 101. My first real lesson in business and startup crash and burn. By early '84, about the same time as the famous Apple Ridley Scott "1984" commercial, we were toast. BCN rattled on with a skeleton crew based at HQ in Riverton WY, by now charging for modem and subscription services.